yatube/requirements.txt
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==2.2.16
django-behaviors==0.5.1
django-debug-toolbar==2.2
django-stubs==1.14.0
django-stubs-ext==0.7.0
djlint==1.19.14
EditorConfig==0.12.3
Faker==12.0.1
flake8==6.0.0
flake8-annotations==3.0.0
flake8-commas==2.1.0
flake8-docstrings==1.7.0
flake8-html==0.4.3
flake8-quotes==3.3.2
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==2.8
iniconfig==2.0.0
Jinja2==3.1.2
jsbeautifier==1.14.7
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.1.2
more-itertools==9.0.0
mypy==1.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.14.0
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.22.0
six==1.14.0
snowballstemmer==2.2.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.6
typing_extensions==4.5.0
urllib3==1.25.11
wcwidth==0.2.6
